requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
qdrant-client
openai
yfinance
beautifulsoup4
requests
lxml
python-dotenv
jinja2
feedparser
